# Fix benchmark script for the case if PR has empty description