<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1089.89<br/>plain u32: 1092.17<br/>aliased double: 943.032<br/>plain double: 947.42 | aliased u32: 1090.74<br/>plain u32: 1088.53<br/>aliased double: 946.05<br/>plain double: 944.889 |
| json-render | String: 8.34541ms<br/>Stringstream: 11.5427ms<br/>Vector: 7.56747ms | String: 8.43579ms<br/>Stringstream: 11.5046ms<br/>Vector: 7.55722ms |
| match_ch | 8.40382ms/req at 82 coordinate<br/>0.102486ms/coordinate | 8.24536ms/req at 82 coordinate<br/>0.100553ms/coordinate |
| match_mld | 7.14559ms/req at 82 coordinate<br/>0.0871413ms/coordinate | 7.08015ms/req at 82 coordinate<br/>0.0863433ms/coordinate |
| packedvector | random write:<br/>std::vector 9663.44 ms<br/>util::packed_vector 78666.5 ms<br/>slowdown: 8.14063<br/>random read:<br/>std::vector 8401.28 ms<br/>util::packed_vector 30825.5 ms<br/>slowdown: 3.66914 | random write:<br/>std::vector 9816.86 ms<br/>util::packed_vector 82068.4 ms<br/>slowdown: 8.35995<br/>random read:<br/>std::vector 11105.9 ms<br/>util::packed_vector 33080.8 ms<br/>slowdown: 2.97867 |
| rtree | 1 result:<br/>207.385ms ->  0.0207385 ms/query<br/>10 results:<br/>242.932ms ->  0.0242932 ms/query | 1 result:<br/>201.319ms ->  0.0201319 ms/query<br/>10 results:<br/>238.111ms ->  0.0238111 ms/query |
<!-- BENCHMARK_RESULTS_END -->